Add files via upload
New version allows work to continue when base names file and/or coast line file are not available.  Essentially, have added if statements and file.exist() function call to control behaviour when these files are not present.  Also added note to the settings related to these data file names (see function runScenario, variables coastFile and baseFile) to indicate that an empty string is the appropriate value when these data files are not available.